delete old unused predictor components
