--- src/app/facility/utility-data/predictor-data/manage-predictors/edit-predictor/edit-predictor.component.css
@@ -1,3 +1,0 @@
-button span {
-    color: white;
-}
