--- src/app/facility/utility-data/predictor-data/manage-predictors/manage-predictors.component.html
@@ -1,5 +1,0 @@
-<h4>
-    Manage Predictors
-</h4>
-<router-outlet></router-outlet>
-
