--- src/app/facility/utility-data/predictor-data/manage-predictors/manage-predictors.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-manage-predictors',
-  templateUrl: './manage-predictors.component.html',
-  styleUrls: ['./manage-predictors.component.css']
-})
-export class ManagePredictorsComponent {
-
-}
