--- src/app/facility/utility-data/predictor-data/manage-predictors/predictors-table/predictors-table.component.css
@@ -1,7 +1,0 @@
-th.is-production{
-    width: 125px;
-}
-
-h5 span.fa.fa-temperature-half{
-    font-size: 17.5px;
-}
