--- src/app/facility/utility-data/predictor-data/manage-predictors/predictors-table/predictors-table.component.html
@@ -1,176 +1,0 @@
-<div class="d-flex w-100 justify-content-end mb-2">
-    <!-- <app-table-items-dropdown></app-table-items-dropdown> -->
-    <div class="ps-2">
-        <button class="btn nav-btn" (click)="uploadData()"><i class="fa fa-upload"></i> Import Data</button>
-    </div>
-    <div class="ps-2">
-        <button class="btn action-btn" (click)="addPredictor()"><i class="fa fa-plus"></i> Add
-            New Predictor</button>
-    </div>
-</div>
-
-<ng-container *ngIf="standardPredictors.length != 0">
-    <h5>Standard</h5>
-    <div class="table-responsive">
-        <table class="table utility-data table-sm table-bordered table-hover" #predictorTable>
-            <thead class="sortable">
-                <tr>
-                    <th>
-                        Name
-                    </th>
-                    <th>
-                        Units
-                    </th>
-                    <th class="is-production">
-                        Is Production?
-                    </th>
-                    <th>Actions</th>
-                </tr>
-            </thead>
-            <tbody class="table-group-divider">
-                <tr *ngFor="let predictor of standardPredictors">
-                    <td>
-                        <a (click)="selectEditPredictor(predictor)">
-                            {{predictor.name}}</a>
-                    </td>
-                    <td>
-                        {{predictor.unit}}
-                    </td>
-                    <td>
-                        <span *ngIf="predictor.production">Yes</span>
-                        <span *ngIf="!predictor.production">No</span>
-                    </td>
-                    <td class="actions">
-                        <div class="btn-group">
-                            <button class="btn btn-sm btn-outline" (click)="selectEditPredictor(predictor)"
-                                title="Edit">
-                                <i class="fa fa-pencil"></i>
-                            </button>
-                            <button class="btn btn-sm btn-outline" (click)="selectDelete(predictor)" title="Delete">
-                                <i class="fa fa-trash"></i>
-                            </button>
-                        </div>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-    </div>
-</ng-container>
-
-
-
-<ng-container *ngIf="degreeDayPredictors.length != 0">
-    <hr *ngIf="standardPredictors.length != 0">
-    <h5><span class="fa fa-temperature-half pe-1"></span>Calculated Degree Days</h5>
-    <div class="alert alert-warning text-center" *ngIf="hasWeatherDataWarning">
-        <span class="fa fa-exclamation-circle"></span> Indicates that one or more predictor entries associated with this
-        predictor was calculated off of weather readings with gaps of twelve hours or more. The data may not be
-        accurate. For closer inspection of that weather station click the <span class="fa fa-thermometer-half"></span>
-        button. It may be best to choose a different weather station
-        to pull data from, use the <a (click)="goToWeatherData()">Weather Data</a> page to explore other stations in the
-        area of this facility.
-    </div>
-    <div class="table-responsive">
-        <table class="table utility-data table-sm table-bordered table-hover" #predictorTable>
-            <thead class="sortable">
-                <tr>
-                    <th>
-                        Name
-                    </th>
-                    <th>
-                        Type
-                    </th>
-                    <th>
-                        Threshold
-                    </th>
-                    <th>
-                        Weather Station
-                    </th>
-                    <th>Actions</th>
-                </tr>
-            </thead>
-            <tbody class="table-group-divider">
-                <tr *ngFor="let predictor of degreeDayPredictors">
-                    <td>
-                        <a (click)="selectEditPredictor(predictor)">
-
-                            {{predictor.name}}</a>
-                    </td>
-                    <td>
-                        <ng-container *ngIf="predictor.weatherDataType == 'HDD'">
-                            Heating
-                        </ng-container>
-                        <ng-container *ngIf="predictor.weatherDataType == 'CDD'">
-                            Cooling
-                        </ng-container>
-                    </td>
-                    <td>
-                        <ng-container *ngIf="predictor.weatherDataType == 'HDD'">
-                            {{predictor.heatingBaseTemperature}} &#8457;
-                        </ng-container>
-                        <ng-container *ngIf="predictor.weatherDataType == 'CDD'">
-                            {{predictor.coolingBaseTemperature}} &#8457;
-                        </ng-container>
-                    </td>
-                    <td>
-                        <span class="fa fa-exclamation-circle" *ngIf="predictor.weatherDataWarning">
-                        </span>
-                        {{predictor.weatherStationName}}
-
-                    </td>
-                    <td class="actions">
-                        <div class="btn-group">
-                            <button class="btn btn-sm btn-outline" (click)="selectEditPredictor(predictor)"
-                                title="Edit">
-                                <i class="fa fa-pencil"></i>
-                            </button>
-                            <button class="btn btn-sm btn-outline" (click)="selectDelete(predictor)" title="Delete">
-                                <i class="fa fa-trash"></i>
-                            </button>
-
-                            <button class="btn btn-sm btn-outline" (click)="viewWeatherData(predictor)"
-                                title="View Weather Data">
-                                <i class="fa fa-temperature-half"></i>
-                            </button>
-                        </div>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-    </div>
-</ng-container>
-
-<ng-container *ngIf="facilityPredictors.length == 0">
-    <div class="alert alert-warning">
-        No predictor data found for this facility. Click "<span class="fa fa-plus"></span> Add New Predictor" button to
-        create your first predictor or the "<span class="fa fa-upload"></span> Import Data" button to use the upload
-        data wizard to upload data from a .xlsx file.
-    </div>
-</ng-container>
-<ng-container *ngIf="predictorEntries.length == 0 && facilityPredictors.length != 0">
-    <div class="alert alert-warning">
-        No entries found for this facility. A predictor entry is required to save these predictors.
-    </div>
-</ng-container>
-
-
-<div [ngClass]="{'windowOverlay': predictorToDelete}"></div>
-<div class="popup" [class.open]="predictorToDelete">
-    <div class="popup-header">
-        Delete {{predictorToDelete?.name}}?
-        <button class="item-right" (click)="cancelDelete()">x</button>
-    </div>
-    <div class="popup-body">
-        <p *ngIf="predictorUsedGroupIds.length == 0">WARNING:
-            Are you sure you want to delete "{{predictorToDelete?.name}}"? This cannot be undone.</p>
-        <p class="alert alert-danger" *ngIf="predictorUsedGroupIds.length != 0">
-            WARNING: This predictor is used in an analysis for this facility. Deleting this predictor will alter this
-            analysis.
-            Are you sure you want to delete "{{predictorToDelete?.name}}"? This cannot be undone.
-        </p>
-    </div>
-    <div class="saveCancel popup-footer text-end">
-        <button class="btn btn-secondary" (click)="cancelDelete()">Cancel</button>
-        <button class="btn btn-danger" (click)="confirmDelete()">Confirm Delete</button>
-    </div>
-</div>
