--- src/app/facility/utility-data/predictor-data/predictor-data.component.css
@@ -1,15 +1,0 @@
-
-.fa.fa-check-circle-o{
-    color: green;
-    font-size: larger;
-}
-
-.copying-table .actions .fa, .copying-table input, .copying-table br{
-    display: none;
-}
-.content-padding{
-    padding: 20px 30px;
-    height: calc(100vh - 290px);
-    overflow-y: auto;
-
-}
