--- src/app/facility/utility-data/predictor-data/predictor-data.component.html
@@ -1,6 +1,0 @@
-<div class="wrapper main-content p-0">
-    <app-predictor-tabs></app-predictor-tabs>
-    <div class="content-padding">
-        <router-outlet></router-outlet>
-    </div>
-</div>
