--- src/app/facility/utility-data/predictor-data/predictor-data.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-predictor-data',
-  templateUrl: './predictor-data.component.html',
-  styleUrls: ['./predictor-data.component.css']
-})
-export class PredictorDataComponent {
-
-}
