--- src/app/facility/utility-data/predictor-data/predictor-entries/edit-predictor-entry/edit-predictor-entry.component.css
@@ -1,3 +1,0 @@
-button span {
-    color: white;
-}
