--- src/app/facility/utility-data/predictor-data/predictor-entries/edit-predictor-entry/edit-predictor-entry.component.html
@@ -1,101 +1,0 @@
-<div class="row">
-    <div class="col-12">
-        <div class="saveCancel item-right">
-            <button class="btn btn-secondary" (click)="cancel()">Cancel</button>
-            <button *ngIf="addOrEdit == 'edit'" class="btn btn-primary" (click)="saveAndQuit()"
-                [disabled]="calculatingDegreeDays">Save</button>
-            <button *ngIf="addOrEdit == 'add'" class="btn btn-primary me-2" (click)="saveAndQuit()"
-                [disabled]="calculatingDegreeDays">Save + Quit</button>
-            <button *ngIf="addOrEdit == 'add'" class="btn btn-primary" (click)="saveAndAddAnother()"
-                [disabled]="calculatingDegreeDays">Save +
-                Add Another</button>
-        </div>
-    </div>
-</div>
-
-<form>
-    <p class="bold">Predictor Entry Data</p>
-    <div class="row">
-        <div class="col-4">
-            <label for="predictorDate">Date</label>
-        </div>
-        <div class="col">
-            <div class="input-group">
-                <input id="predictorDate" class="form-control" type="month"
-                    [ngModel]="predictorEntry.date | date:'yyyy-MM'" (ngModelChange)="setDate($event)" name="date">
-            </div>
-        </div>
-    </div>
-    <br>
-
-    <div *ngFor="let predictor of predictorEntry.predictors">
-        <div class="row">
-            <div class="col-4">
-                <label [for]="'predictor-' + predictor.id">
-                    <span class="fa fa-temperature-half" *ngIf="predictor.predictorType == 'Weather'"></span>
-                    {{predictor.name}}
-                    <ng-container *ngIf="predictor.predictorType == 'Weather'">
-                        <br>
-                        <span class="small block">({{predictor.weatherDataType}},
-                            <ng-container *ngIf="predictor.weatherDataType == 'HDD'">
-                                {{predictor.heatingBaseTemperature}} &#8457;
-                            </ng-container>
-                            <ng-container *ngIf="predictor.weatherDataType == 'CDD'">
-                                {{predictor.coolingBaseTemperature}} &#8457;
-                            </ng-container>)</span>
-                    </ng-container>
-                </label>
-            </div>
-            <div class="col">
-                <div class="input-group mb-2">
-                    <ng-container
-                        *ngIf="predictor.predictorType == 'Weather' && (predictor.weatherDataWarning || predictor.weatherOverride)">
-                        <span class="input-group-text">
-                            <span class="fa fa-exclamation-circle" *ngIf="predictor.weatherDataWarning"></span>
-                            <span class="fa fa-asterisk" *ngIf="predictor.weatherOverride"></span>
-                        </span>
-                    </ng-container>
-
-                    <input [id]="'predictor-' + predictor.id" class="form-control" type="number"
-                        [(ngModel)]="predictor.amount" name="{{'predictorAmount'+predictor.id}}"
-                        onfocus="this.select();"
-                        [disabled]="predictor.predictorType == 'Weather' && !predictor.weatherOverride"
-                        (input)="setChanged()">
-                    <span class="input-group-text" *ngIf="predictor.unit">{{predictor.unit}}</span>
-                    <ng-container
-                        *ngIf="predictor.predictorType == 'Weather' && (predictor.weatherDataWarning || predictor.weatherOverride)">
-                        <span class="input-group-text p-0">
-                            <button class="btn btn-dark btn-sm h-100" *ngIf="!predictor.weatherOverride"
-                                (click)="setWeatherManually()">
-                                <span class="fa fa-edit"></span>
-                            </button>
-                            <button class="btn btn-dark btn-sm h-100" *ngIf="predictor.weatherOverride"
-                                (click)="revertManualWeatherData()">
-                                <span class="fa fa-calculator"></span>
-                            </button>
-                        </span>
-                    </ng-container>
-                </div>
-            </div>
-        </div>
-    </div>
-</form>
-
-<div *ngIf="hasWeatherDataWarning" class="alert alert-warning text-center">
-    <span class="fa fa-exclamation-circle"></span> Indicates this value is calculated off of weather data with gaps of
-    twelve hours or more. The resulting value may not be accurate. It may be best to choose a different weather station
-    to pull data from, use the <a (click)="goToWeatherData()">Weather Data</a> page to explore other stations in the
-    area of this facility. <a (click)="setWeatherManually()">Set the weather data manually.</a>
-</div>
-
-<div *ngIf="hasWeatherOverride" class="alert alert-warning text-center">
-    <span class="fa fa-asterisk"></span> Indicates this weather data entry has been directly set and the degree day
-    calculation has been skipped. <a (click)="revertManualWeatherData()">Calculate Weather Data</a>
-</div>
-<div *ngIf="hasWeatherData" class="alert alert-info text-center">
-    Degree day predictors are denoted with a <span class="fa fa-temperature-half"></span> icon and are automatically
-    calculated.
-</div>
-
-<app-calculating-spinner *ngIf="calculatingDegreeDays"
-    [message]="'Calculating Degree Days...'"></app-calculating-spinner>
