--- src/app/facility/utility-data/predictor-data/predictor-entries/predictor-entries.component.html
@@ -1,4 +1,0 @@
-<h4>
-    Predictor Entries
-</h4>
-<router-outlet></router-outlet>
