--- src/app/facility/utility-data/predictor-data/predictor-entries/predictor-entries.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-predictor-entries',
-  templateUrl: './predictor-entries.component.html',
-  styleUrls: ['./predictor-entries.component.css']
-})
-export class PredictorEntriesComponent {
-
-  constructor() { }
-}
