--- src/app/facility/utility-data/predictor-data/predictor-tabs/predictor-tabs.component.html
@@ -1,20 +1,0 @@
-<ul class="nav nav-tabs">
-    <li class="nav-item">
-        <a class="nav-link" routerLink="manage" [routerLinkActive]="['active']">
-            <i class="fa fa-gears"></i>
-            Manage Predictors
-        </a>
-    </li>
-    <li class="nav-item" [ngClass]="{'disabled': facilityPredictors.length == 0}">
-        <a class="nav-link" routerLink="entries" [routerLinkActive]="['active']">
-            <i class="fa fa-table"></i>
-            Predictor Entries
-        </a>
-    </li>
-    <!-- <li class="nav-item">
-        <a class="nav-link" routerLink="degree-days" [routerLinkActive]="['active']">
-            <i class="fa fa-table"></i>
-            Degree Days
-        </a>
-    </li> -->
-</ul>
