+++ src/app/facility/utility-data/predictors/predictors-data/predictors-data-table/predictors-data-table.component.ts
@@ -235,11 +235,11 @@ export class PredictorsDataTableComponent {
   }
 
   goToWeatherData() {
-    // let facility: IdbFacility = this.facilityDbService.selectedFacility.getValue();
-    // this.weatherDataService.selectedFacility = facility;
-    // let selectedFacility: IdbFacility = this.facilityDbService.selectedFacility.getValue();
-    // this.weatherDataService.selectedFacility = selectedFacility;
-    // this.weatherDataService.zipCode = selectedFacility.zip;
-    // this.router.navigateByUrl('/weather-data');
+    let facility: IdbFacility = this.facilityDbService.selectedFacility.getValue();
+    this.weatherDataService.selectedFacility = facility;
+    let selectedFacility: IdbFacility = this.facilityDbService.selectedFacility.getValue();
+    this.weatherDataService.selectedFacility = selectedFacility;
+    this.weatherDataService.zipCode = selectedFacility.zip;
+    this.router.navigateByUrl('/weather-data');
   }
 }
